backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
exceptiongroup==1.1.3
fastapi==0.115.6
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
googlesearch-python==1.2.5
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
MarkupSafe==2.1.5
nltk==3.8.2
outcome==1.2.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
PyMuPDF==1.25.1
pyparsing==3.2.1
pypdfium2==4.30.1
PySocks==1.7.1
python-certifi-win32==1.6.1
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
redis==5.2.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
selenium==4.12.0
setuptools-scm==8.1.0
shellingham==1.5.4
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.41.3
textblob==0.18.0.post0
tqdm==4.66.5
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.4
utils==1.0.2
uvicorn==0.32.1
vine==5.1.0
watchfiles==1.0.3
websockets==14.1
Werkzeug==3.0.3
wrapt==1.17.2
wsproto==1.2.0
